feat(nfts): Create useBunnySpecialLottery hook and massive refactor
